[OpenAI] Async iterables to disposable and iterable readable streams (#28079)
### Packages impacted by this PR
@azure/core-sse, @azure/openai

### Issues associated with this PR
Fixes https://github.com/Azure/azure-sdk-for-js/issues/27959

### Describe the problem that is addressed by this PR
This PR updates the SSE operations to return a readable stream instead
of an async iterable, which addresses the following two issues:
- The issue above describes a scenario where exiting early from the loop
leaves the socket not completely drained which in turn causes the
application to hang until the socket timeouts and closes. This PR
returns streams instead which will be canceled once the loop exits
early, and that cancellation will prompt drainage of the socket.
Moreover, the stream is disposable so the socket will be drained once
the stream goes out of scope but that requires the variable declaration
to be by `await using`
- The output stream enables pipe chaining scenarios such as
text-to-speech

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
Alternatively, the linked issue could be addressed by adding the logic
for draining the socket in the definition of the `return` method on the
AsyncIterable interface. @bterlson mentioned this could also be
addressed by adding the drainage logic in the `finally` clause around
the internal async iteration implementation which I needed to do anyway
for the browser case because at the moment, [only ff supports async
iteration protocol for
streams](https://streams.spec.whatwg.org/#rs-asynciterator).

### Are there test cases added in this PR? _(If not, why?)_
Yes.

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)