[core] call onResponse in as*Stream methods (#28111)
### Packages impacted by this PR
@azure-rest/core-client and @azure/openai

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/28093

### Describe the problem that is addressed by this PR
The `onResponse` callback option is not called in `streamCompletions`
and `streamChatCompletions`. This is because the `asNodeStream` and
`asBrowserStream` methods in the internal client don't call them.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
This PR addresses the issue by simply calling the onResponse call back
in the `asNodeStream` and `asBrowserStream` methods. I don't think we
want to _not_ call onResponse there.

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)