[OpenAI & Core] Fix whisper responseFormat behavior (#28148)
### Packages impacted by this PR
@azure/openai & @azure/core-rest-pipeline

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/28145

### Describe the problem that is addressed by this PR

The `responseFormat` parameter is included in a form data request. If
the customer does not specify it, the parameter is added to the form
data request with a value of undefined. However, the form data policy
attempts to treat this value as a File object.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

To avoid sending an incorrect parameter, the fix is to not send it if it
is not set by the customer. Additionally, the core should verify that
the form data value is defined and is an object before proceeding to
treat it as such.

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)