[core] Default to `application/octet-stream` content type for multipart blobs (#28199)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`

### Issues associated with this PR

- Fixes #28081

### Describe the problem that is addressed by this PR

According to the `multipart/form-data` specification, binary content
should have a default Content-Type of `application/octet-stream` if none
is provided. Not supplying a `Content-Type` value implies that the
content is text-based with a default encoding.

This PR corrects behavior so that form fields with values of type `Blob`
and `File` have their content type set to `application/octet-stream` in
the multipart request if no other content type is specified.

### Are there test cases added in this PR? _(If not, why?)_

Fixed existing test to conform to corrected behavior.