[core] Enforce ordering of formData and multipart policies (#28224)
### Packages impacted by this PR

@azure/core-rest-pipeline
@azure/ts-http-runtime

### Describe the problem that is addressed by this PR

This PR makes two minor changes to the core packages, neither should
result in behavior change at runtime.

- Moves setting the requests' multipartBody outside the loop to avoid
setting it repeatedly. The end result is the same, since we return early if there
is no form data.
- Explicitly declares an ordering between multipartPolicy and
formDataPolicy to ensure the two are always sequenced correctly. 
This happens implicitly today but will now be enforced.

Both of these changes were spin-offs of changes in #28083

### Are there test cases added in this PR? _(If not, why?)_

There are no behavior changes to test - these should be considered pure
refactors

### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/28083/

### Checklists
- [x] Added impacted package name to the issue description.
- [x] Does this PR need any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here.)_
- [x] Added a changelog (if necessary).