[core-amqp, service-bus] Correct timeToLive calculation to use absolute expiry time (#28178)
Following the .NET fix
https://github.com/Azure/azure-sdk-for-net/pull/40916

### Packages impacted by this PR
`@azure/core-amqp`, `@azure/service-bus`

### Issues associated with this PR
#28121 

### Describe the problem that is addressed by this PR

When a message is received, if the absolute expiry time was set on the
AMQP message, the TTL should be set using the difference between the
absolute expiry time and the message creation time. This works around
the limitation of the AMQP spec where a value larger than 49 days in
milliseconds cannot be sent over the wire as the field is a uint32. Also
currently JS SDK throws an error if a TTL value of larger than 49 days
in milliseconds value is set on messages to send.

Both core-amqp and service-bus are fixed because service-bus has its own
`fromRheaMessage` and `toRheaMessage` to enable configuring JSON body
parsing behavior and Date conversion.