Feature/ru based capping for fetch requests (#28273)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR


### Describe the problem that is addressed by this PR
In the Node SDK, when performing a cross-partition query and using the
FetchNext() method, the SDK will not return results until it has
retrieved the number of items specified by the maxItemCount in the query
options. This behaviour, while ensuring complete data retrieval, can
potentially lead to high consumption of Request Units (RU's) if the
dataset spans numerous partitions.
To address this issue, it is imperative to implement a mechanism that
allows for control over the number of RUs utilized per fetch operation,
whether it's via fetchNext() or fetchAll().

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
To enable Query operations to adhere to a Request Unit (RU) cap, we can
offer clients the option to specify the RU cap within the operation
options of fetch functions. When this cap is exceeded, clients will
receive an error notification.
For more details check this design docs:

https://microsoftapc-my.sharepoint.com/:w:/g/personal/abkolant_microsoft_com/EeWSsuBdil9PiSJ30xYGuK4B7J79hF0WwC3ZeuVtCb_fRg?e=s2MAuG

https://microsoftapc-my.sharepoint.com/:w:/g/personal/mkhandelwal_microsoft_com/EXLW5MXJagVHhLb0tjgpWl4BCBpm3CGHSzqWqRRlXAf5MQ?e=n4a43V

### Are there test cases added in this PR? (If not, why?)
yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Manik Khandelwal <mkhandelwal@microsoft.com>