[Identity] Initialize public client application before API calls for msal-browser - IBC (#28292)
### Packages impacted by this PR
@azure/identity

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/28215

### Describe the problem that is addressed by this PR
- The public client application needs to be initialized explicitly with
the new version of msal-browser. This affects the browser-side's
Interactive Browser Credential which throws the
`unitialized_public_client` error from @azure/msal-browser when trying
to authenticate with the getToken() call.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
- We can either call the initialize() method on the public application's
instance OR we can call the `createPublicClientApplication()` method for
initializing the public client application, like we are doing in this PR

### Are there test cases added in this PR? _(If not, why?)_
- No, will open a separate PR for test cases and samples for IBC
credentials. As this focuses on ONLY the changes needed for the bug fix.

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Added a changelog (if necessary)