Reverting previous changes of version 5.1.0 (#28331)
### Packages impacted by this PR

@azure/eventgrid

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

This reverts the changes of version 5.1.0. The system events introduced
in the previous version are still not ready. They will be introduced
later.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No special design considerations.

### Are there test cases added in this PR? _(If not, why?)_

No test cases required.

### Provide a list of related PRs _(if any)_

NA

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

NA

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec Please review and approve the PR. Thanks