[identity] Remove MsalBaseUtils class (#28283)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

N/A 

### Describe the problem that is addressed by this PR

As we start to cleanup and refactor Identity, one of the low hanging
fruits we can remove right now is the MsalBaseUtils class
which all the MsalFlows inherit from. By standardizing here we can start
to peel away some of the inheritance based indirection
and move towards composition of smaller building blocks.

I tried to minimize the number of changes and make this a pure refactor,
as 1:1 to the existing behavior as possible while we
work out the larger refactor of the Msal inheritance based flows.