[OpenAI] Fix streaming (#28351)
### Packages impacted by this PR
@azure/openai

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/28342

### Describe the problem that is addressed by this PR
In `TransformStream`, the transform function shouldn't call
`controller.terminate()` unless it is expected that an error will be
thrown in the code that is consuming the writable stream. This PR
removes that call. Thanks @xirzec for the investigation!

UPDATE: Since the transformer stream is no longer ending, the idea of a
disposition operation is irrelevant. This is because the underlying
readable stream remains perpetually linked to the transformer stream,
making it impossible to cancel a stream that's locked. Now that I think
about it, the concept of disposition seems unnecessary in this case to
begin with. Once you're finished with the stream, there's no real need
for disposal so I deleted it. If a user wishes to cancel the stream,
they should use the standard method of calling the cancel API.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
We don't test on Bun but we have plenty chat completions tests.

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)