identity: DefaultAzureCredential chain refactor (#28264)
### Packages impacted by this PR

@azure/identity 

### Issues associated with this PR

N/A

### Describe the problem that is addressed by this PR

As noted in #25024, this constructor pattern presents a few challenges:

1. Credentials that throw in their constructor cause the
DefaultAzureCredential to throw, even if they won't end up being used
2. Requiring credentials to be flexible in their constructor in order to
support being added to a chained credential requires either overloading
the constructor or having a single constructor accept an options bag
rather than being able to be explicit about what we're passing in
3. Overloaded constructors to support the `Default*Credential` scenario
adds code and complexity to a credential that would otherwise be
contained in the `Default*Credential`

This PR refactors the default credential to use a set of factory
functions instead of subclasses. It also introduces a "no-op" that can
be used to safely store a reason the credential was skipped. Finally, it
guards against any errors being thrown from the underlying credential.

All-in-all, a pure refactor that adds admittedly little benefit for any
existing credentials (we should be able to simplify them, but the public
constructors make that a breaking change so the ship has sailed) but
creates a path forward and a pattern for adding new credentials to the
chain.

In the future we may be able to simplify ManagedIdentityCredential and
WorkloadIdentityCredential but that is a ways away.

### Provide a list of related PRs _(if any)_
https://github.com/Azure/azure-sdk-for-js/pull/25024

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [x] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)