identity: replace SNI tests with unit tests (#28361)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

N/A

### Describe the problem that is addressed by this PR

Our live tests have been failing since November. One of the failures
relates to an expired certificate that is used to validate SNI. 

Chatting with the identity feature crew it sounds like SNI is a
first-party only feature requiring the kind of cert that we may not be able to renew.

An alternative proposal involved making sure that we send the expected
x5c header to the token request's assertion when sendCertificateChain is
true.

We have a set of internal tests that this could be added to. The
simplest test could check that when we parse the certificate we 
correctly set the x5c field when sendCertificateChain is true, 
and omit it when false. The rest flows naturally from there.