[core] Avoid use of web streams in multipart/form-data (#28294)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`

### Issues associated with this PR

- #28102
- Fixes https://github.com/Azure/autorest.typescript/issues/2177

### Describe the problem that is addressed by this PR


- Added private "raw content" property to `createFile`-created `File`s
to work around a couple of Node bugs (one where `Blob#stream` doesn't
work in certain versions of Node 20, and another where
`Readable.fromWeb` doesn't work in certain versions of Node 18). The
"raw" property is used to get the original stream instead of the
`stream` method. This property is used when accessing the stream so we
don't have to convert to/from web streams, cirvumventing the two issues.
- In browser, multipartPolicy now concatenates `Blob` objects using the
`Blob` constructor instead of concatenating `ReadableStream`s. This
works around the fact that `ReadableStream` is not a supported request
body type when using HTTP/1.1 in the browser (see #28102).
- We are still using the browser-mapped formDataPolicy for now, but this
should unblock us returning to the isomorphic multipartPolicy-based
formDataPolicy if it provides a benefit in future.
- See this doc for some interesting reading on the design of the `Blob`
API as it is implemented by Chromium:
https://chromium.googlesource.com/chromium/src/+/HEAD/storage/browser/blob/README.md.
Notably, large `Blob`s may be put on disk when created, and creating a
new `Blob` by concatenating other `Blob`s does not involve an expensive
copy.