Number lookup GA (#27711)
GA implementation of the number lookup feature.  This commit should cover the vast majority of the work needed, and is updated to incorporate feedback from the stewardship board.  These changes are going into a branch to wait until we're ready to do our final bug bashing and eventually merging and shipping the feature.

-----

### Packages impacted by this PR
communication-phone-numbers

### Issues associated with this PR
https://skype.visualstudio.com/SPOOL/_workitems/edit/3431961/

### Describe the problem that is addressed by this PR
Updating the SDK to GA implementation

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
This design was approved by the Review Board in our API design review.
The original discussion on design choices is below:

> The main consideration for this design is how to structure the output
data and how to accept input from the caller that they would like to
include additional paid fields in the response. Three fields are grouped
financially (e.g. requesting one costs the same as requesting all 3),
but not conceptually. So we would prefer to have the user specify a
single option (in this case an "IncludeOperatorDetails" option) and get
back all 3 fields. There are 2 other options, each with their own
pros/cons:
> 
> Alternative 1: Use the 'expand' query parameter
> One way to design this is to use the 'expand' parameter and ask the
caller to include all 3 fields in their request. The benefit of this
approach is that it adheres more closely to the azure standard, but the
downside is that if a customer specifies only one or two of the fields
they will not be getting their money's worth for the query (e.g. paying
for OperatorDetails but only getting IsoCountryCode).
> 
> Alternative 2: Restructure the data
> We could reshape the return type to group the three fields that cost
the same amount. This would allow us to both use the 'expand' parameter
and also have a clean story for billing. However, the downside is that
our data structure wouldn't quite make sense as these fields are only
grouped because of cost, not because they are logically connected.
Logically they should all be top-level properties. This leaks
implementation details into the API design and locks us into a specific
billing structure that is tied to our third party data providers

### Are there test cases added in this PR? _(If not, why?)_
yes

### Provide a list of related PRs _(if any)_
Retaining links, but these PRs are closed as we have chosen the design
in this PR.

> Alternative 1: Use the 'expand' parameter:
https://github.com/Azure/azure-sdk-for-js/pull/27712
> Alternative 2: Restructure the data:
https://github.com/Azure/azure-sdk-for-js/pull/27713

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)