[EngSys] ensure dev dependency `@types/mocha` when depending on `mocha`
Our tests use the type `Mocha.Context` when they need to pass the test context
to test recorder. However, without a dev dependency on `@types/mocha`
compilation may fail when recorder constructor is changed from taking a
`Mocha.Test` to taking an interface that is compatible.

This change ensures the dev depedency on `@types/mocha` too if there's a dev
dependency on `mocha`.

***NO_CI***