[core-rest-pipeline] Re-enable multipartPolicy-based formDataPolicy in browser (#28457)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`

### Issues associated with this PR

- https://github.com/Azure/azure-sdk-for-js/issues/28102
- https://github.com/Azure/autorest.typescript/issues/2178
- https://github.com/Azure/azure-sdk-for-js/issues/28355

### Describe the problem that is addressed by this PR

This PR re-enables the `multipartPolicy`-based `formDataPolicy` in the
browser. The browser implementation of `multipartPolicy`, as of
https://github.com/Azure/azure-sdk-for-js/issues/28294, relies on
creating `Blob`s and concatenating them into a larger `Blob`. This means
that files from disk (provided by the user via an `<input type="file">`
element) are not loaded into memory. Avoiding that was the main
motivation to move back to the browser-implementation-based policy, but
it's no longer an issue.

I believe we're at the point where bringing this back is worthwhile. It
gives us finer control over the wire format of the multipart request,
which we want because:

- It is a stepping stone to enable support for `application/json` parts
in `multipart/form-data` requests. Using the `FormData`-based policy for
this is impossible because we can't set a content type without also
setting a file name using `FormData` (leaving the file name unspecified
results in the browser setting the file name to "blob" which isn't what
we want)
- It allows for customization of the multipart boundary, required for
recorded tests to work

Doing this should fix a bug where files created using
`createFileFromStream` don't work in the browser. We could fix this in
the `FormData`-based policy, but if we do this we won't have to :)

As a follow-up to this PR I plan to add a flag to `multipartPolicy` to
re-enable streaming support for customers who want to stream content
(other than on-disk files, which are already supported) and are
confident that their environment supports HTTP/2.

### Are there test cases added in this PR?

I added further test cases to `formDataPolicy` to ensure that arrays are
supported properly, that combining file fields and text fields works,
and also to ensure `createFileFromStream` works properly in browser.