[core] allows choosing Fetch or XMLHttpRequest for browser HttpClient
Currently we don't expose the concrete types for our HttpClients. For browsers
we returns a `FetchHttpClient` from `createDefaultHttpClient`. However, some
times a `XhrHttpClient` is desired, for example, because it has better support
on upload progress reporting. But since `XhrHttpClient` is not exposed,
customers cannot use it.

This PR provides an option to `createDefaultHttpClient()`, allowing choosing the
underlying API.