Adding new events for EG Version 5.2.0 (#28513)
### Packages impacted by this PR

@azure/eventgrid

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

The EventGrid Service Team has added 14 new events. The SDK must be
regenerated to include the code changes related to these 14 new events.
The SDK minor version has been updated.

  - `Microsoft.Storage.StorageTaskAssignmentQueued`
  - `Microsoft.Storage.StorageTaskAssignmentCompleted`
  - `Microsoft.AVS.ScriptExecutionStarted`
  - `Microsoft.AVS.ScriptExecutionFinished`
  - `Microsoft.AVS.ScriptExecutionCancelled`
  - `Microsoft.AVS.ScriptExecutionFailed`
  - `Microsoft.AVS.PrivateCloudUpdating`
  - `Microsoft.AVS.PrivateCloudUpdated`
  - `Microsoft.AVS.PrivateCloudFailed`
  - `Microsoft.AVS.ClusterCreated`
  - `Microsoft.AVS.ClusterDeleted`
  - `Microsoft.AVS.ClusterUpdating`
  - `Microsoft.AVS.ClusterUpdated`
  - `Microsoft.AVS.ClusterFailed`
 
### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

There are no specific/complex design scenarios for this task. It is a
straightforward regenerate and some standard changes to the custom layer
of the code.

### Are there test cases added in this PR? _(If not, why?)_

No. This item is standard and we need not add test cases for every new
events. The existing cases would be sufficient.

### Provide a list of related PRs _(if any)_

- https://github.com/Azure/azure-sdk-for-js/pull/27384 (This is the PR
that adds similar events to the SDK in the 4.15.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/26939 (This is the PR
that adds similar events to the SDK in the 4.14.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/26020 (This is the PR
that adds similar events to the SDK in the 4.13.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/28176 (This is the PR
that adds similar events to the SDK in the 5.1.0 release)


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

`autorest --typescript swagger\README.md`

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)