[dev-tool] Customization v2 prototype (#28503)
### Packages impacted by this PR

- `@azure-tools/dev-tool`

### Describe the problem addressed by this PR

New concept for customization tool I want to check in so we can test
whether it works better for our needs than the existing tool.

As a high-level overview, this tool leverages Git and its diff/merge
capabilities to take handwritten modifications to the old generated code
and apply it to new generated code after regenerating. The user will
resolve any merge conflicts if present after the fact.