[EngSys] add a workaround to load ESM tests for Mocha  (#28556)
Currently most of our '*.js' tests and sources under dist-esm/ are
transformed by `esm` package (now outdated and archived) before being
feeded to Mocha because when a package.json has `"type":
"commonjs"` thus Mocha treats `.js` files as CommonJS.

This PR takes another approach by telling Mocha that the loaded modules
under dist-esm/ are ESM via a custom loader (`esm4mocha.mjs`).
This is done mainly in two parts:

1. During resolving, this loader checks the path of resolved modules and
if it includes `"dist-esm/"` we set the format of the resolved module to
`"module"`.
2. Also in ESM the full extension is required for relative modules in
import/export declarations, so a transformation is applied
to add `".js"` extension to relative imports/exports on the fly when
loading the module.

ServiceBus integration-test:node script has been updated to use this
approach. For now, the dev-tool command is unrolled and the old
workaround is replaced by the new one:

```diff
- -r ../../../common/tools/esm-workaround -r esm
+ --loader=../../../common/tools/esm4mocha.mjs
```

In the future, the dev-tool command would be updated to use the new
workaround for the majority of our packages.