[keyvault] Prepare for 7.5 GA release (#28416)
### Packages impacted by this PR

- `@azure/keyvault-admin`
- `@azure/keyvault-certificates`
- `@azure/keyvault-keys`
- `@azure/keyvault-secrets`

### Issues associated with this PR

- Fixes #28455

### Describe the problem that is addressed by this PR

Preparing for release of the new Key Vault packages supporting the new
7.5 service version.
- Rerecorded tests for all packages
- Updated constants and types to reflect new default service version

This GA will introduce the following features:
- keyvault-admin: Allow using a user-assigned Managed Identity to access
the Storage account when performing backup and restore operations.
- keyvault-certificates: Add a new property `x509ThumbprintString` to
`CertificateProperties`. This property is a hex string representation of
the existing `x509Thumbprint` property, provided for convenience and for
interoperability with ARM.
- keyvault-secrets: Add new property `hsmPlatform` to `KeyProperties`.

These things have changed since the beta release:
- Based on architect feedback, updated backup/restore operations to
introduce an overload when using Managed Identity instead of making the
`sasToken` parameter optional. This avoids the user having to pass
`undefined` to `sasToken` if the customer wishes to set request options.