[core] disable tshy selfLink
The feature is to support any modules inside of the package to be able to import
from the package itself.  Currently we only ever need that for samples
development and for that we add "paths" in tsconfig.json to point to the package's
src/index.ts. We haven't needed to import a sub path.

This PR disables selfLink for now until there's a need. Because with it enabled,
tshy creates symlinks under the dist/ directory which broken build cache.