[core-tracing] support sharing state between CJS and ESM (#28631)
### Packages impacted by this PR

@azure/core-tracing

### Issues associated with this PR

Resolves #28619

### Describe the problem that is addressed by this PR

If you are exporting both CommonJS and ESM forms of a package, then it
is possible for both versions to be loaded at run-time.
However, the CommonJS build is a different module from the ESM build,
and thus a different thing from the point of view of the
JavaScript interpreter in Node.js.

> https://github.com/isaacs/tshy/blob/main/README.md#dual-package-hazards

tshy handles this by building programs into separate folders and treats
"dual module hazards" as a fact of life. 

One of the hazards of dual-modules is shared module-global state.

In core-tracing we have a module-global instrumenter that is used for
hooking into and lighting up OpenTelemetry based instrumentation. In order to
ensure it works in this dual-package world we must use one of multiple-recommended
workarounds.

In this case, the tshy documentation provides a solution to this with a
well-documented path forward. This is what is implemented here.

Please refer to
https://github.com/isaacs/tshy/blob/main/README.md#module-local-state
for added context