[identity] Prepare cache persistence plugin for release (#28659)
### Packages impacted by this PR

@azure/identity-cache-persistence

### Issues associated with this PR

Resolves #28608

### Describe the problem that is addressed by this PR

@azure/identity-cache-persistence was never released with msal-node 2.x
support.
As a result, it does not support node 20.

Further, the [released
version](https://deps.dev/npm/%40azure%2Fidentity-cache-persistence/1.0.0/dependencies)
still depends on an alpha version of msal-node-extensions.

This PR preps the release of 1.1.0 with fixes to both of these.

### Are there test cases added in this PR? _(If not, why?)_

There are no tests in this package unfortunately. I will validate
manually
against our nightly build and add a tracking issue to add tests.