Pre Push Hook to check and fix code format (#28500)
### Packages impacted by this PR
No specific package is impacted. This PR is to ensure that the code is
formatted before checking the code to a github repository.

### Issues associated with this PR
No

### Describe the problem that is addressed by this PR
In our repository we have several projects. When we make changes to them
and start the PR process, the build and analyze process starts. During
this process, the code formatter checks and reports any errors. In an
ideal world, it works great. But, in reality, I have found these errors
only after 30 - 45 minutes after creating the PR. This has occurred many
times and it would be great if there is a solution that can report
formatting issue before I create the PR.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
A few things are considered in the current design:
1. This solution must help the developer and not create another hurdle.
So, If the developer wants to do multiple commits locally, without
worrying about formatting, he should be able to do it. (So, pre-commit
hook is not used)
2. The solution must work across multiple projects and multiple commits.
3. Direct Git hooks are the ideal solution. But, unfortunately, the
hooks cannot be distributed to the team, since git ignores everything
under `.git` folder.
4. So, I have used the rush hooks directory. So, this will be available
to all the team members who clone the repository. Now, they have 2
options. A) They could copy the hook to their `.git\hooks\` folder. B) A
better recommended solution is to execute the following command: `rush
install` which will automatically handle the linking for them

### Are there test cases added in this PR? _(If not, why?)_
No. This is a pre-push git hook

### Provide a list of related PRs _(if any)_
None

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
None

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@xirzec @deyaaeldeen @jeremymeng @joheredi @maorleger Please review and
approve the PR. Thanks